tipo del atributo cantidad de apuestas cambiado

+++ BBDD/Ejercicio Apuesta deportivas.sql
@@ -59,7 +59,7 @@ CREATE TABLE Apuestas(
 	--------------------------------------------------------------------------
 	cuota decimal(5,2) not null
 		constraint ck_Apuestas_cuota check(cuota > 1),
-	cantidad int not null,
+	cantidad money not null,
 		--constraint ck_Apuestas_cantidad check(cantidad > 0),
 	tipo char(1) not null
 		constraint ck_Apuestas_tipo check(tipo in ('1','2','3')),
